adding simple authentication with key
