+++ .env
@@ -2,6 +2,7 @@ MONGODB_URI=mongodb://admin:password@192.168.122.195:27017/?test?authSource=admi
 MONGO_DATABASE="recipes"
 INITIAL_CREATEDB="false"
 REDIS_HOST="192.168.122.195"
+X-API-KEY="tosaGvPTYpn74cZgsFRwuQ=="
 ##docker run this way
 #docker run -d --name mongodb -v /home/dani/mongodb/data:/data/db -e  MONGO_INITDB_ROOT_USERNAME=admin -e MONGO_INITDB_ROOT_PASSWORD=pass -p 27017:27017 mongo:4.4.18
 #docker run -it --rm \
